refactor: consolidate duplicate mask_password implementations
Extract mask_password function to shared utils module to follow DRY principle.
All files now import from utils.password_utils instead of having duplicate
implementations.

- Create utils/password_utils.py with shared mask_password function
- Update api_service.py to use shared function
- Update test_masked_password.py to use shared function
- Update examples/masked_password_example.py to use shared function
- Update verify_email_password.py to use shared function

This addresses the code review feedback from PR #38 about
consolidating duplicate implementations.
