+++ examples/masked_password_example.py
@@ -2,21 +2,11 @@
 """
 Example showing how masked passwords appear in the API response
 """
+import sys
+import os
+sys.path.insert(0, os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
 
-
-def mask_password(password):
-    """Mask a password showing only first 2 and last 2 characters."""
-    if not password:
-        return None
-
-    if len(password) <= 4:
-        return "*" * len(password)
-
-    first_two = password[:2]
-    last_two = password[-2:]
-    middle_stars = "*" * (len(password) - 4)
-
-    return f"{first_two}{middle_stars}{last_two}"
+from utils.password_utils import mask_password
 
 
 def simulate_api_response():
